+++ include/Texture.hpp
@@ -1,72 +1,15 @@
 #ifndef RAYLIB_CPP_INCLUDE_TEXTURE_HPP_
 #define RAYLIB_CPP_INCLUDE_TEXTURE_HPP_
 
-#include <string>
-
-#include "./raylib.hpp"
-#include "./raylib-cpp-utils.hpp"
-#include "./Vector2.hpp"
-#include "./Material.hpp"
-#include "./RaylibException.hpp"
+#include "./TextureUnmanaged.hpp"
 
 namespace raylib {
 /**
  * Texture type
  */
-class Texture : public ::Texture {
+class Texture : public TextureUnmanaged {
  public:
-    /**
-     * Default texture constructor.
-     */
-    Texture() : ::Texture{0, 0, 0, 0, 0} {
-        // Nothing.
-    }
-
-    /**
-     * Move/Create a texture structure manually.
-     */
-    Texture(unsigned int id,
-            int width, int height,
-            int mipmaps = 1,
-            int format = PIXELFORMAT_UNCOMPRESSED_R8G8B8A8)
-            : ::Texture{id, width, height, mipmaps, format} {
-        // Nothing.
-    }
-
-    /**
-     * Creates a texture object based on the given Texture struct data.
-     */
-    Texture(const ::Texture& texture) :
-            ::Texture{texture.id, texture.width, texture.height, texture.mipmaps, texture.format} {}
-
-    /**
-     * Creates a texture from the given Image.
-     *
-     * @throws raylib::RaylibException Throws if failed to create the texture from the given image.
-     */
-    Texture(const ::Image& image) {
-        Load(image);
-    }
-
-    /**
-     * Load cubemap from image, multiple image cubemap layouts supported.
-     *
-     * @throws raylib::RaylibException Throws if failed to create the texture from the given cubemap.
-     *
-     * @see LoadTextureCubemap()
-     */
-    Texture(const ::Image& image, int layout) {
-        Load(image, layout);
-    }
-
-    /**
-     * Load texture from file into GPU memory (VRAM)
-     *
-     * @throws raylib::RaylibException Throws if failed to create the texture from the given file.
-     */
-    Texture(const std::string& fileName) {
-        Load(fileName);
-    }
+    using TextureUnmanaged::TextureUnmanaged;
 
     Texture(const Texture&) = delete;
 
@@ -84,17 +27,6 @@ class Texture : public ::Texture {
         Unload();
     }
 
-    GETTERSETTER(unsigned int, Id, id)
-    GETTERSETTER(int, Width, width)
-    GETTERSETTER(int, Height, height)
-    GETTERSETTER(int, Mipmaps, mipmaps)
-    GETTERSETTER(int, Format, format)
-
-    Texture& operator=(const ::Texture& texture) {
-        set(texture);
-        return *this;
-    }
-
     Texture& operator=(const Texture&) = delete;
 
     Texture& operator=(Texture&& other) noexcept {
@@ -102,7 +34,6 @@ class Texture : public ::Texture {
             return *this;
         }
 
-        Unload();
         set(other);
 
         other.id = 0;
@@ -112,330 +43,6 @@ class Texture : public ::Texture {
         other.format = 0;
 
         return *this;
-    }
-
-    /**
-     * Retrieve the width and height of the texture.
-     */
-    inline ::Vector2 GetSize() const {
-        return {static_cast<float>(width), static_cast<float>(height)};
-    }
-
-    /**
-     * Load texture from image data
-     */
-    void Load(const ::Image& image) {
-        set(::LoadTextureFromImage(image));
-        if (!IsReady()) {
-            throw RaylibException("Failed to load Texture from Image");
-        }
-    }
-
-    /**
-     * Load cubemap from image, multiple image cubemap layouts supported
-     */
-    void Load(const ::Image& image, int layoutType) {
-        set(::LoadTextureCubemap(image, layoutType));
-        if (!IsReady()) {
-            throw RaylibException("Failed to load Texture from Cubemap");
-        }
-    }
-
-    /**
-     * Load texture from file into GPU memory (VRAM)
-     */
-    void Load(const std::string& fileName) {
-        set(::LoadTexture(fileName.c_str()));
-        if (!IsReady()) {
-            throw RaylibException("Failed to load Texture from file: " + fileName);
-        }
-    }
-
-    /**
-     * Unload texture from GPU memory (VRAM)
-     */
-    inline void Unload() {
-        ::UnloadTexture(*this);
-    }
-
-    /**
-     * Update GPU texture with new data
-     */
-    inline Texture& Update(const void *pixels) {
-        ::UpdateTexture(*this, pixels);
-        return *this;
-    }
-
-    /**
-     * Update GPU texture rectangle with new data
-     */
-    inline Texture& Update(::Rectangle rec, const void *pixels) {
-        UpdateTextureRec(*this, rec, pixels);
-        return *this;
-    }
-
-    /**
-     * Get pixel data from GPU texture and return an Image
-     */
-    inline ::Image GetData() const {
-        return ::LoadImageFromTexture(*this);
-    }
-
-    /**
-     * Get pixel data from GPU texture and return an Image
-     */
-    inline operator raylib::Image() {
-        return GetData();
-    }
-
-    /**
-     * Generate GPU mipmaps for a texture
-     */
-    inline Texture& GenMipmaps() {
-        ::GenTextureMipmaps(this);
-        return *this;
-    }
-
-    /**
-     * Set texture scaling filter mode
-     */
-    inline Texture& SetFilter(int filterMode) {
-        ::SetTextureFilter(*this, filterMode);
-        return *this;
-    }
-
-    /**
-     * Set texture wrapping mode
-     */
-    inline Texture& SetWrap(int wrapMode) {
-        ::SetTextureWrap(*this, wrapMode);
-        return *this;
-    }
-
-    /**
-     * Draw a Texture2D
-     *
-     * @see ::DrawTexture()
-     */
-    inline Texture& Draw(int posX = 0, int posY = 0, ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTexture(*this, posX, posY, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a Texture2D with position defined as Vector2
-     *
-     * @see ::DrawTextureV()
-     */
-    inline Texture& Draw(::Vector2 position, ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureV(*this, position, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a Texture2D with extended parameters
-     *
-     * @see ::DrawTextureEx()
-     */
-    inline Texture& Draw(::Vector2 position, float rotation, float scale = 1.0f,
-            ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureEx(*this, position, rotation, scale, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a part of a texture defined by a rectangle
-     *
-     * @see ::DrawTextureRec()
-     */
-    inline Texture& Draw(::Rectangle sourceRec, ::Vector2 position = {0, 0},
-            ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureRec(*this, sourceRec, position, tint);
-        return *this;
-    }
-
-    /**
-     * Draw texture quad with tiling and offset parameters
-     *
-     * @see ::DrawTextureQuad()
-     */
-    inline Texture& Draw(::Vector2 tiling, ::Vector2 offset, ::Rectangle quad,
-            ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureQuad(*this, tiling, offset, quad, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a part of a texture defined by a rectangle with 'pro' parameters
-     *
-     * @see ::DrawTexturePro()
-     */
-    inline Texture& Draw(::Rectangle sourceRec, ::Rectangle destRec, ::Vector2 origin = {0, 0},
-            float rotation = 0, ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTexturePro(*this, sourceRec, destRec, origin, rotation, tint);
-        return *this;
-    }
-
-    /**
-     * Draws a texture (or part of it) that stretches or shrinks nicely
-     *
-     * @see ::DrawTextureNPatch()
-     */
-    inline Texture& Draw(::NPatchInfo nPatchInfo, ::Rectangle destRec, ::Vector2 origin = {0, 0},
-            float rotation = 0, ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureNPatch(*this, nPatchInfo, destRec, origin, rotation, tint);
-        return *this;
-    }
-
-    /**
-     * Draw part of a texture (defined by a rectangle) with rotation and scale tiled into dest.
-     *
-     * @see ::DrawTextureTiled()
-     */
-    inline Texture& DrawTiled(::Rectangle sourceRec, ::Rectangle destRec, ::Vector2 origin = {0, 0},
-            float rotation = 0, float scale = 1, Color tint = {255, 255, 255, 255}) {
-        ::DrawTextureTiled(*this, sourceRec, destRec, origin, rotation, scale, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a textured polygon
-     *
-     * @see ::DrawTexturePoly()
-     */
-    inline Texture& DrawPoly(::Vector2 center, ::Vector2 *points,
-            ::Vector2 *texcoords, int pointsCount,
-            ::Color tint = {255, 255, 255, 255}) {
-        ::DrawTexturePoly(*this, center, points, texcoords, pointsCount, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a billboard texture
-     *
-     * @see ::DrawBillboard()
-     */
-    inline Texture& DrawBillboard(const ::Camera& camera,
-            ::Vector3 position, float size,
-            ::Color tint = {255, 255, 255, 255}) {
-        ::DrawBillboard(camera, *this, position, size, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a billboard texture defined by source
-     *
-     * @see ::DrawBillboardRec()
-     */
-    inline Texture& DrawBillboard(const ::Camera& camera,
-            ::Rectangle source, ::Vector3 position, ::Vector2 size,
-            ::Color tint = {255, 255, 255, 255}) {
-        DrawBillboardRec(camera, *this, source, position, size, tint);
-        return *this;
-    }
-
-    /**
-     * Draw a billboard texture defined by source and rotation
-     *
-     * @see ::DrawBillboardPro()
-     */
-    inline Texture& DrawBillboard(const ::Camera& camera,
-            ::Rectangle source, Vector3 position,
-            ::Vector3 up, Vector2 size, Vector2 origin, float rotation = 0.0f,
-            ::Color tint = {255, 255, 255, 255}) {
-        DrawBillboardPro(camera, *this, source, position, up, size, origin, rotation, tint);
-        return *this;
-    }
-
-    /**
-     * Draw cube textured
-     *
-     * @see ::DrawCubeTexture()
-     */
-    inline Texture& DrawCube(::Vector3 position, float width, float height, float length,
-            ::Color color = {255, 255, 255, 255}) {
-        ::DrawCubeTexture(*this, position, width, height, length, color);
-        return *this;
-    }
-
-    /**
-     * Draw cube textured, with dimensions
-     *
-     * @see ::DrawCubeTexture()
-     */
-    inline Texture& DrawCube(::Vector3 position, ::Vector3 dimensions,
-            ::Color color = {255, 255, 255, 255}) {
-        ::DrawCubeTexture(*this, position, dimensions.x, dimensions.y, dimensions.z, color);
-        return *this;
-    }
-
-    /**
-     * Draw cube with a region of a texture
-     *
-     * @see ::DrawCubeTextureRec()
-     */
-    inline Texture& DrawCube(::Rectangle source, ::Vector3 position, float width, float height, float length,
-            ::Color color = {255, 255, 255, 255}) {
-        ::DrawCubeTextureRec(*this, source, position, width, height, length, color);
-        return *this;
-    }
-
-    /**
-     * Draw cube with a region of a texture, with dimensions
-     *
-     * @see ::DrawCubeTextureRec()
-     */
-    inline Texture& DrawCube(::Rectangle source, ::Vector3 position, ::Vector3 dimensions,
-            ::Color color = {255, 255, 255, 255}) {
-        ::DrawCubeTextureRec(*this, source, position, dimensions.x, dimensions.y, dimensions.z, color);
-        return *this;
-    }
-
-    /**
-     * Set texture for a material map type (MAP_DIFFUSE, MAP_SPECULAR...)
-     */
-    inline Texture& SetMaterial(::Material *material, int mapType = MATERIAL_MAP_NORMAL) {
-        ::SetMaterialTexture(material, mapType, *this);
-        return *this;
-    }
-
-    inline Texture& SetMaterial(const ::Material& material, int mapType = MATERIAL_MAP_NORMAL) {
-        ::SetMaterialTexture((::Material*)(&material), mapType, *this);
-        return *this;
-    }
-
-    /**
-     * Set texture and rectangle to be used on shapes drawing.
-     */
-    inline Texture& SetShapes(const ::Rectangle& source) {
-        ::SetShapesTexture(*this, source);
-        return *this;
-    }
-
-    /**
-     * Set shader uniform value for texture (sampler2d)
-     */
-    inline Texture& SetShaderValue(const ::Shader& shader, int locIndex) {
-        ::SetShaderValueTexture(shader, locIndex, *this);
-        return *this;
-    }
-
-    /**
-     * Determines whether or not the Texture has been loaded and is ready.
-     *
-     * @return True or false depending on whether the Texture has data.
-     */
-    bool IsReady() const {
-        return id != 0;
-    }
-
- private:
-    void set(const ::Texture& texture) {
-        id = texture.id;
-        width = texture.width;
-        height = texture.height;
-        mipmaps = texture.mipmaps;
-        format = texture.format;
     }
 };
